Reorganized hops module
