+++ src/java/growthcraft/hops/ClientProxy.java
@@ -1,5 +1,7 @@
 package growthcraft.hops;
 
+import growthcraft.hops.renderer.RenderHops;
+
 import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class ClientProxy extends CommonProxy
