+++ src/java/growthcraft/hops/item/ItemHopSeeds.java
@@ -1,6 +1,10 @@
-package growthcraft.hops;
+package growthcraft.hops.item;
 
 import growthcraft.core.GrowthCraftCore;
+import growthcraft.hops.GrowthCraftHops;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.player.EntityPlayer;
@@ -8,8 +12,6 @@ import net.minecraft.init.Blocks;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemHopSeeds extends Item
 {
