+++ src/java/growthcraft/hops/item/ItemHops.java
@@ -1,10 +1,11 @@
-package growthcraft.hops;
+package growthcraft.hops.item;
 
 import growthcraft.core.GrowthCraftCore;
-import net.minecraft.client.renderer.texture.IIconRegister;
-import net.minecraft.item.Item;
+
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import net.minecraft.client.renderer.texture.IIconRegister;
+import net.minecraft.item.Item;
 
 public class ItemHops extends Item
 {
