+++ src/java/growthcraft/hops/renderer/RenderHops.java
@@ -1,13 +1,15 @@
-package growthcraft.hops;
+package growthcraft.hops.renderer;
 
 import growthcraft.core.Utils;
+import growthcraft.hops.block.BlockHops;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderHops implements ISimpleBlockRenderingHandler
 {
